% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConservationTweets}
\alias{ConservationTweets}
\title{Tweets mention palm oil and plastic polluion}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5902 rows and 88 columns.}
\source{
Twitter.com using rTweet and twitter API.
}
\usage{
data(ConservationTweets)
}
\description{
Data containing tweets with keywords "palm oil plastic pollution", presented as a dataframe with all
tweet relevant information.
}
\examples{
df <- ConservationTweets
tokenizedTweets <- tokenizeTweet(df,1)
}
\keyword{Conservarion}
\keyword{Palm}
\keyword{Twitter}
\keyword{oil}
