% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{dispFreq}
\alias{dispFreq}
\title{dispFreq}
\usage{
dispFreq(tokenizedTweetVarName = TidyTweetsTokenized)
}
\arguments{
\item{tokenizedTweetVarName}{Leave blank unless the name of the tokenized Tweets data frame was changed}
}
\value{
Displays the most frequent words used.
}
\description{
Displays the most common words/pair/triplesof words.
}
\examples{
dispFreq() dispFreq(TweetsTokenized)
}
