% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodellingFunctions.R
\name{estTopics}
\alias{estTopics}
\title{estTopics}
\usage{
estTopics(tweetDataFrameTokenized, minTopics = 2, maxTopics = 30,
  stepNum = 1)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}

\item{minTopics}{Lower bound of topics to be optimised over.}

\item{maxTopics}{Upper bound of topics to be optimised over.}

\item{stepNum}{Optimisation step amount. (1 will take longer but be more accurate)}
}
\value{
Optimised range of topics for specific data frame variable.
}
\description{
Produce a plot with using 4 optimisationg techniques to determine range of optimal topic numbers for the LDA model.
}
\examples{
estTopics(tweetDataFrameTokenized,2,30,1
}
