% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{loadTweets}
\alias{loadTweets}
\title{loadTweets}
\usage{
loadTweets(fileName)
}
\arguments{
\item{fileName}{File name containing streamed Tweets. Format: "myfilename.json"}
}
\value{
Data frame object containing streamed tweets and relevant information.
}
\description{
Load .json file containing streamed tweets into a data frame object.
}
\examples{
tweetDataFrame <- loadTweets("myfilename.json")
}
