% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotBigrams}
\alias{plotBigrams}
\title{plotBigrams}
\usage{
plotBigrams(mostFreqWords = TweetFreq, minWordFreq = 5)
}
\arguments{
\item{mostFreqWords}{Name of variable containing the frequency of words used, leave blank if unchanged}

\item{minWordFreq}{Minimum frequency Bigram to appear on the cloud plot}
}
\value{
Directional cloud plot displaying connectivity between Bigrams
}
\description{
Plot a cloud plot showing how bigrams connect together and the frequency of bigrams
}
\examples{
plotBigrams(,10) plotBigrams() plotBigrams(mostFrequentlyUsedWords,)
}
