% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotCompBar}
\alias{plotCompBar}
\title{plotCompBar}
\usage{
plotCompBar(mostFreqWords = TweetFreq, maxWords = 10,
  minWordFreq = 1)
}
\arguments{
\item{mostFreqWords}{Name of variable containing the frequency of words used, leave blank if unchanged}

\item{maxWords}{n number of most commonly used words to be displayed}

\item{minWordFreq}{Minimum frequency of word to appear on the cloud plot}
}
\value{
Bar chart containing n most frequent positive and negative words in the Tweet data set
}
\description{
Plot a bar chart of the n most negative and positive words in the Tweet data set
}
\examples{
plotCompBar() plotCompBar(,10,5) plotCompBar(mostFrequentlyUsedWords,100,100)
}
