% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotCompCloud}
\alias{plotCompCloud}
\title{plotCompCloud}
\usage{
plotCompCloud(tweetDataFrameTokenized, maxWords = 50, minWordFreq = 1)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}

\item{maxWords}{Maximum number of words to appear on the plot.}

\item{minWordFreq}{Minimum frequency of word to appear on the plot}
}
\value{
Cloud plot containing n most frequent positive and negative words in the Tweet data set
}
\description{
Plot a comparison cloud of the n most negative and positive words in the Tweet data set
}
\examples{
plotCompMap(tweetDataFrameTokenized) plotCompMap(tweetDataFrameTokenized,10,5)
}
