% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotCompMap}
\alias{plotCompMap}
\title{plotCompMap}
\usage{
plotCompMap(mostFreqWords = TweetFreq, maxWords = 50,
  minWordFreq = 1)
}
\arguments{
\item{mostFreqWords}{Name of variable containing the frequency of words used, leave blank if unchanged}

\item{maxWords}{n number of most commonly used words to be displayed}

\item{minWordFreq}{Minimum frequency of word to appear on the cloud plot}
}
\value{
Cloud plot containing n most frequent positive and negative words in the Tweet data set
}
\description{
Plot a comparison cloud of the n most negative and positive words in the Tweet data set
}
\examples{
plotCompMap() plotCompMap(,10,5) plotCompMap(mostFrequentlyUsedWords,100,100)
}
