% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotFreqBar}
\alias{plotFreqBar}
\title{plotFreqBar}
\usage{
plotFreqBar(mostFreqWords = TweetFreq, numOfWords = 10)
}
\arguments{
\item{mostFreqWords}{Name of variable containing the frequency of words used, leave blank if unchanged}

\item{numOfWords}{Top n words to be plotted, default is 10.}
}
\value{
Top n most frequent words used in the Tweet dataset.
}
\description{
Plot a bar chart of the n most frequent words in the Tweet data set
}
\examples{
plotFreqBar(,100) plotFreqBar() plotFreqBar(mostFrequentlyUsedWords,5)
}
