% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotFreqCloud}
\alias{plotFreqCloud}
\title{plotFreqCloud}
\usage{
plotFreqCloud(mostFreqWords = TweetFreq, maxWords = 70,
  minWordFreq = 100)
}
\arguments{
\item{mostFreqWords}{Name of variable containing the frequency of words used, leave blank if unchanged}

\item{maxWords}{n number of most commonly used words to be displayed}

\item{minWordFreq}{Minimum frequency of word to appear on the cloud plot}
}
\value{
Cloud plot containing n number of most frequent words within the Tweet dataset
}
\description{
Plot a cloud plot containing the most frequently used words within the Tweet data set
}
\examples{
plotFreqCloud() plotFreqCloud(,10,5) plotFreqCloud(mostFrequentlyUsedWords,100,100)
}
