% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotTopicTerms}
\alias{plotTopicTerms}
\title{plotTopicTerms}
\usage{
plotTopicTerms(ldaModel, topN = 10)
}
\arguments{
\item{ldaModel}{lda model with numerous topics, output of tweetLDA().}

\item{topN}{number of top terms to be plotted for each topic}
}
\value{
Plot of most common terms within each topic
}
\description{
Plot most terms associated with each topic
}
\examples{
plotTopicTerms(tweetDataFrameTokenized,10)
}
