% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{removeStopWords}
\alias{removeStopWords}
\title{removeStopWords}
\usage{
removeStopWords(tidyTweetsTokenizedVarName = TidyTweetsTokenized,
  extraStopWords = NULL)
}
\arguments{
\item{tidyTweetsTokenizedVarName}{the variable name of the data frame containing the tokenized tweets. Leave blank if unchanged}

\item{extraStopWords}{If there are extra words to exclude from the analysis enter them here, format: "c("word1","word2")}
}
\value{
Tokenized Tweets data frame excluding stop words.
}
\description{
Call and remove stop words from the tokenized tweet data set, stop words include common words that are not interesting for analysis (the,and, a, ...)
}
\examples{
removeStopWords() removeStopWords(,c("tree","garden")) removeStopWords(tokenizedTweetsNew,"pear")
}
