% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{streamTweets}
\alias{streamTweets}
\title{streamTweets}
\usage{
streamTweets(searchWords, searchPeriod, fileName = "streamedTweets.json")
}
\arguments{
\item{searchWords}{Words or phrases to be included in tweets, format: "word1,word2,a phrase"}

\item{searchPeriod}{Number of seconds that Tweets will be collected for}

\item{fileName}{File name of the results of the function, format: "myfilename.json"}
}
\value{
File in local directory with Tweets containing desired words and/or phrases.
}
\description{
Search for Tweets containing entered words/phrases over designated time period and save with desired file name in file directory.
}
\examples{
streamTweets("tree,bushes,apple orchard",60,"treeTweets.json")
}
