% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{tokenFreq}
\alias{tokenFreq}
\title{tokenFreq}
\usage{
tokenFreq(tweetDataFrameTokenized)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}
}
\value{
Displays the most frequent words used within Tweet data.
}
\description{
Displays the most common words/pair/triplesof words.
}
\examples{
mostCommonTokens <- tokenFreq(tweetDataFrameTokenized)
}
