% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{tokenizeTweets}
\alias{tokenizeTweets}
\title{tokenizeTweets}
\usage{
tokenizeTweets(tidyTweetVarName = TidyTweets, tokenNum = 1)
}
\arguments{
\item{tidyTweetVarName}{data frame variable containing imported streamed tweets, leave blank if name was not changed.}

\item{tokenNum}{1 - splits into singular words, 2 - splits into pairs of words, 3 - splits into triples of words.}
}
\value{
Data frame object containing all Tweets split into n number of words.
}
\description{
Splits contents of tweets into words/pairs of words/ triples of words depending on the tokenization number.
}
\examples{
tokenizeTweets(Tweets,3) tokenizeTweets(,2) tokenizeTweets()
}
