% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{tweetContext}
\alias{tweetContext}
\title{tweetContext}
\usage{
tweetContext(wordReq, tidyTweetsVarName = TidyTweets)
}
\arguments{
\item{WordReq}{The word or phrase wanted for context, format: "word1"}

\item{TidyTweetsVarName}{Name of the variable containing the imported tweets, if unchanged then leave blank.}
}
\value{
Full length Tweet containing the word or phrase searched for.
}
\description{
If uncertain about the context of a popular word, the full length of the tweet will be shown to give context.
}
\examples{
tweetContext("pineapple") tweetContext("pineapple tree") tweetContext("pineapple",myImportedTweets)
}
