% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{plotBigrams}
\alias{plotBigrams}
\title{plotBigrams}
\usage{
plotBigrams(tweetDataFrameTokenized, minWordFreq = 5)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}

\item{minWordFreq}{Minimum frequency of word to appear on the plot.}
}
\value{
Directional cloud plot displaying connectivity between Bigrams
}
\description{
Plot a cloud plot showing how bigrams connect together and the frequency of bigrams
}
\examples{
plotBigrams(tweetDataFrameTokenized,10) plotBigrams(tweetDataFrameTokenized)
}
