% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{plotCompBar}
\alias{plotCompBar}
\title{plotCompBar}
\usage{
plotCompBar(tweetDataFrameTokenized, maxWords = 10, minWordFreq = 1)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}

\item{maxWords}{Maximum number of words to appear on the plot.}

\item{minWordFreq}{Minimum frequency of word to appear on the plot.}
}
\value{
Bar chart containing n most frequent positive and negative words in the Tweet data set
}
\description{
Plot a bar chart of the n most negative and positive words in the Tweet data set
}
\examples{
plotCompBar(tweetDataFrameTokenized) plotCompBar(tweetDataFrameTokenized,10,5)
}
