% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{plotFreqCloud}
\alias{plotFreqCloud}
\title{plotFreqCloud}
\usage{
plotFreqCloud(tweetDataFrameTokenized, maxWords = 70,
  minWordFreq = 100)
}
\arguments{
\item{tweetDataFrameTokenized}{Dataframe object containing word tokens from tweets (Output of tokenizeTweets())}

\item{maxWords}{Maximum number of words to appear on the plot.}

\item{minWordFreq}{Minimum frequency of word to appear on the plot}
}
\value{
Cloud plot containing n number of most frequent words within the Tweet dataset
}
\description{
Plot a cloud plot containing the most frequently used words within the Tweet data set
}
\examples{
plotFreqCloud(tweetDataFrameTokenized) plotFreqCloud(tweetDataFrameTokenized,10,5)
}
