% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageRef.R
\name{tweetContext}
\alias{tweetContext}
\title{tweetContext}
\usage{
tweetContext(tweetDataFrame, wordReq)
}
\arguments{
\item{tweetDataFrame}{Name of the dataframe object containing all tweets (output of loadTweets())}

\item{WordReq}{The word or phrase wanted for context, format: "word1"}
}
\value{
Full length Tweet containing the word or phrase searched for.
}
\description{
If uncertain about the context of a popular word, the full length of the tweet will be shown to give context.
}
\examples{
tweetContext(tweetDataFrame,"pineapple") tweetContext(tweetDataFrame,"pineapple tree")
}
